feature/Movies2.0_001_movies Fixed naming
