+++ app/src/test/java/com/veselovvv/movies20/movies/presentation/MoviesUiTest.kt
@@ -1,7 +1,7 @@
 package com.veselovvv.movies20.movies.presentation
 
 import com.veselovvv.movies20.core.Order
-import com.veselovvv.movies20.movies.presentation.FakeMovieDomainToUiMapper.Companion.MOVIE_MAP
+import com.veselovvv.movies20.movies.presentation.FakeMovieDomainToUiMapper.Companion.MOVIE_MAP_UI
 import com.veselovvv.movies20.movies.presentation.FakeMoviesCommunication.Companion.MAP
 import org.junit.Before
 import org.junit.Test
@@ -56,7 +56,7 @@ class MoviesUiTest {
         )
         communication.checkMapCalledCount(1)
         movieMapper.checkMapCalledCount(1)
-        order.check(listOf(MOVIE_MAP, MAP))
+        order.check(listOf(MOVIE_MAP_UI, MAP))
 
         ui = MoviesUi.Success(movies = listOf(), movieMapper = movieMapper)
         ui.map(mapper = communication)
@@ -64,7 +64,7 @@ class MoviesUiTest {
         communication.checkList(listOf<MovieUi>(MovieUi.NoResults))
         communication.checkMapCalledCount(2)
         movieMapper.checkMapCalledCount(1)
-        order.check(listOf(MOVIE_MAP, MAP, MAP))
+        order.check(listOf(MOVIE_MAP_UI, MAP, MAP))
     }
 
     @Test
