+++ app/src/test/java/com/veselovvv/movies20/movies/presentation/MoviesViewModelTest.kt
@@ -4,8 +4,6 @@ import com.veselovvv.movies20.core.Order
 import com.veselovvv.movies20.movies.presentation.FakeMovieCache.Base.Companion.SAVE_MOVIE_INFO
 import com.veselovvv.movies20.movies.presentation.FakeMoviesCommunication.Companion.MAP
 import com.veselovvv.movies20.movies.presentation.MoviesViewModelTest.FakeFetchMoviesUseCase.Companion.FETCH_EXECUTE
-import com.veselovvv.movies20.movies.presentation.MoviesViewModelTest.FakeMoviesDomainToUiMapper.Companion.MAP_MOVIES_FAIL
-import com.veselovvv.movies20.movies.presentation.MoviesViewModelTest.FakeMoviesDomainToUiMapper.Companion.MAP_MOVIES_SUCCESS
 import com.veselovvv.movies20.movies.presentation.MoviesViewModelTest.FakeSearchMoviesUseCase.Companion.SEARCH_EXECUTE
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.runBlocking
@@ -279,54 +277,6 @@ class MoviesViewModelTest {
                             movieMapper = BaseMovieDataToDomainMapper()
                         )
                 } else MoviesDomain.Fail(error = ErrorType.GENERIC_ERROR)
-            }
-        }
-    }
-
-    interface FakeMoviesDomainToUiMapper : MoviesDomainToUiMapper {
-        companion object {
-            const val MAP_MOVIES_SUCCESS = "FakeMoviesDomainToUiMapper#mapsuccess"
-            const val MAP_MOVIES_FAIL = "FakeMoviesDomainToUiMapper#mapfail"
-        }
-
-        fun checkSuccessMapCalledCount(count: Int)
-        fun checkFailMapCalledCount(count: Int)
-
-        class Base(private val order: Order) : FakeMoviesDomainToUiMapper {
-            private var successMapCalledCount = 0
-            private var failMapCalledCount = 0
-
-            override fun checkSuccessMapCalledCount(count: Int) {
-                assertEquals(count, successMapCalledCount)
-            }
-
-            override fun checkFailMapCalledCount(count: Int) {
-                assertEquals(count, failMapCalledCount)
-            }
-
-            override fun map(movies: List<MovieDomain>): MoviesUi {
-                successMapCalledCount++
-                order.add(MAP_MOVIES_SUCCESS)
-                return MoviesUi.Success(movies, BaseMovieDomainToUiMapper())
-            }
-
-            override fun map(error: ErrorType): MoviesUi {
-                failMapCalledCount++
-                order.add(MAP_MOVIES_FAIL)
-
-                return MoviesUi.Fail(
-                    when (error) {
-                        ErrorType.NO_CONNECTION -> NO_CONNECTION_MESSAGE
-                        ErrorType.SERVICE_UNAVAILABLE -> SERVICE_UNAVAILABLE_MESSAGE
-                        else -> SOMETHING_WENT_WRONG
-                    }
-                )
-            }
-
-            companion object {
-                private const val NO_CONNECTION_MESSAGE = "No connection. Please try again!"
-                private const val SERVICE_UNAVAILABLE_MESSAGE = "Service unavailable. Please try again!"
-                private const val SOMETHING_WENT_WRONG = "Something went wrong. Please try again!"
             }
         }
     }
